+++ IsekaiMod/Content/Features/IsekaiProtagonist/IsekaiAuraSelection.cs
@@ -73,19 +73,14 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist {
             });
 
             const string DarkAuraName = "Dark Aura";
-            const string DarkAuraDesc = "Enemies within 40 feet take a –4 penalty on attack {g|Encyclopedia:Dice}rolls{/g}, AC, and saving throws.";
-            const string DarkAuraDescBuff = "This creature has a –4 penalty on attack {g|Encyclopedia:Dice}rolls{/g}, AC, and saving throws.";
+            const string DarkAuraDesc = "Enemies within 40 feet take a –4 penalty on AC and saving throws.";
+            const string DarkAuraDescBuff = "This creature has a –4 penalty on AC and saving throws.";
             var Icon_Dark_Aura = AssetLoader.LoadInternal(IsekaiContext, "Features", "ICON_DARK_AURA.png");
             var DarkAuraBuff = ThingsNotHandledByTTTCore.CreateBuff("DarkAuraBuff", bp => {
                 bp.SetName(IsekaiContext, DarkAuraName);
                 bp.SetDescription(IsekaiContext, DarkAuraDescBuff);
                 bp.IsClassFeature = true;
                 bp.m_Icon = Icon_Dark_Aura;
-                bp.AddComponent<AddStatBonus>(c => {
-                    c.Descriptor = ModifierDescriptor.Penalty;
-                    c.Stat = StatType.AdditionalAttackBonus;
-                    c.Value = -4;
-                });
                 bp.AddComponent<AddStatBonus>(c => {
                     c.Descriptor = ModifierDescriptor.Penalty;
                     c.Stat = StatType.AC;